engine: extract esm/mercator_tile.js
